feat(balances): refactor/rename exports and add documentation
