+++ packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/balances/index.ts
@@ -1,11 +1,4 @@
-import * as custodialBalanceSelectors from './custodial.selectors'
-import * as fiatBalanceSelectors from './fiat.selectors'
-import * as nonCustodialBalanceSelectors from './non-custodial.selectors'
-import * as totalBalanceSelectors from './total.selectors'
-
-export default {
-  ...custodialBalanceSelectors,
-  ...fiatBalanceSelectors,
-  ...nonCustodialBalanceSelectors,
-  ...totalBalanceSelectors
-}
+export * from './custodial.selectors'
+export * from './fiat.selectors'
+export * from './non-custodial.selectors'
+export * from './total.selectors'
